requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
discord.py==2.3.2
